s892

+++ utils/gnn_loss.py
@@ -60,7 +60,7 @@ class gnn_loss_noatt(nn.Module):
             # pred_hb = F.softmax(input=pred_hb, dim=1)
             # loss_hb.append(lovasz_softmax_flat(*flatten_probas(pred_hb, targets[1], self.ignore_index),
             #                           only_present=self.only_present))
-            loss_hb.append(self.criterion(pred_hb, targets[1]))
+            loss_hb.append(self.criterion(pred_hb, targets[1].long()))
         loss_hb = sum(loss_hb)
         #half seg loss final
         pred1 = F.interpolate(input=preds[1][-1], size=(h, w), mode='bilinear', align_corners=True)
@@ -78,7 +78,7 @@ class gnn_loss_noatt(nn.Module):
             # pred_fb = F.softmax(input=pred_fb, dim=1)
             # loss_fb.append(lovasz_softmax_flat(*flatten_probas(pred_fb, targets[2], self.ignore_index),
             #                           only_present=self.only_present))
-            loss_fb.append(self.criterion(pred_fb, targets[2]))
+            loss_fb.append(self.criterion(pred_fb, targets[2].long()))
         loss_fb = sum(loss_fb)
         #full seg loss final
         pred2 = F.interpolate(input=preds[2][-1], size=(h, w), mode='bilinear', align_corners=True)
